requirements.txt
requests
python-dotenv

psycopg2-binary
snowflake-connector-python[pandas]
sqlalchemy

pandas
apache-airflow